requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
beautifulsoup4
